backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
ultralytics>=8.1.0
opencv-python>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
